rename package rule to datasource
