+++ shardingsphere-scaling/shardingsphere-scaling-core/src/main/java/org/apache/shardingsphere/scaling/core/config/datasource/DataSourceConfiguration.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.scaling.core.config.rule;
+package org.apache.shardingsphere.scaling.core.config.datasource;
 
 import org.apache.shardingsphere.infra.database.type.DatabaseType;
